refactor(backend): add comment headers, delete helpers
Almost all helpers were wrappers for commands.
