+++ src-tauri/src/call/mod.rs
@@ -1,3 +1,1 @@
-// call/mod.rs
-// Ojos Project
 pub mod commands;
