+++ src-tauri/src/contacts/mod.rs
@@ -1,4 +1,3 @@
 pub mod commands;
-pub mod helper;
 pub mod schema;
 pub mod structs;
