+++ src-tauri/src/contacts/structs.rs
@@ -1,3 +1,7 @@
+// File:     contacts/structs.rs
+// Purpose:  Structs for the Contacts tool.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use serde::{Deserialize, Serialize};
 
 /// # `Contact` struct
