+++ src-tauri/src/helpers.rs
@@ -1,5 +1,12 @@
+// File:     helpers.rs
+// Purpose:  Tasks that need to be completed often, such as timestamps.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use rusqlite::Connection;
-use std::time::{SystemTime, UNIX_EPOCH};
+use std::{
+    path::PathBuf,
+    time::{SystemTime, UNIX_EPOCH},
+};
 use tauri::{AppHandle, Manager};
 use uuid::Uuid;
 
@@ -17,4 +24,8 @@ pub fn stamp() -> (i64, String) {
 pub fn db_connect(app: &AppHandle) -> Connection {
     Connection::open(app.path().app_data_dir().unwrap().join("iris.db"))
         .expect("Connection to iris.db failed.")
+}
+
+pub fn data_dir(app: &AppHandle) -> PathBuf {
+    app.path().app_data_dir().unwrap()
 }
