+++ src-tauri/src/pro/commands.rs
@@ -1,7 +1,12 @@
+// File:     pro/commands.rs
+// Purpose:  Commands for the Patient Reported Outcomes tool.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use super::structs::PatientReportedOutcome;
-use super::{helper, structs::ProQuestion};
-use rusqlite::Connection;
-use tauri::{AppHandle, Manager};
+use super::structs::ProQuestion;
+use crate::helpers::{db_connect, stamp};
+use rusqlite::named_params;
+use tauri::AppHandle;
 
 /// # `add_pro` Command
 ///
@@ -14,7 +19,17 @@ use tauri::{AppHandle, Manager};
 /// ```
 #[tauri::command(rename_all = "snake_case")]
 pub fn add_pros(app: AppHandle, pros: Vec<(String, i32)>) {
-    helper::add_pros(app, pros);
+    let conn = db_connect(&app);
+    let (timestamp, uuid) = stamp();
+
+    for pro_tuple in pros {
+        // INSERT OR IGNORE will skip an insertion if a primary or unique constraint
+        // is failed.
+        conn.execute(
+            "INSERT INTO patient_recorded_outcome (id, recorded_date, question, response) VALUES (:uuid, :recorded_date, :question, :response)", 
+            named_params! {":uuid": uuid, ":recorded_date": timestamp, ":question": pro_tuple.0, ":response": i64::try_from(pro_tuple.1).unwrap()}
+        ).expect("Inserting into patient_recorded_outcome failed.");
+    }
 } // MAKE SURE pro tuples are formatted with question first and response second!
 
 /// # `get_all_pros` Command
@@ -31,13 +46,34 @@ pub fn add_pros(app: AppHandle, pros: Vec<(String, i32)>) {
 /// ```
 #[tauri::command(rename_all = "snake_case")]
 pub fn get_all_pros(app: AppHandle) -> Vec<PatientReportedOutcome> {
-    helper::get_all_pros(app)
+    let conn = db_connect(&app);
+
+    let mut pros_list: Vec<PatientReportedOutcome> = vec![];
+
+    let mut statement = conn
+        .prepare("SELECT * FROM patient_recorded_outcome")
+        .expect("Could not fetch PROs from database.");
+
+    let matched_pros = statement
+        .query_map([], |row| {
+            Ok(PatientReportedOutcome {
+                id: row.get(0)?,
+                recorded_date: row.get(1)?,
+                question: row.get(2)?,
+                response: row.get(3)?,
+            })
+        })
+        .unwrap();
+
+    for pro in matched_pros {
+        pros_list.push(pro.unwrap());
+    }
+    pros_list
 }
 
 #[tauri::command(rename_all = "snake_case")]
 pub fn get_pro_questions(app: AppHandle) -> Vec<ProQuestion> {
-    let path = app.path().app_data_dir().unwrap();
-    let conn = Connection::open(path.join("iris.db")).unwrap();
+    let conn = db_connect(&app);
 
     let mut stmt = conn
         .prepare("SELECT * FROM pro_question")
