+++ src-tauri/src/pro/structs.rs
@@ -1,3 +1,7 @@
+// File:     pro/structs.rs
+// Purpose:  Structs for the Patient Reported Outcomes tool.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
