+++ src-tauri/src/settings/mod.rs
@@ -1,3 +1,2 @@
 pub mod commands;
-pub mod config;
 pub mod structs;
