+++ src-tauri/src/updater/commands.rs
@@ -1,51 +1,12 @@
+// File:     updater/commands.rs
+// Purpose:  The menu at the top of the app for desktop.
+// Authors:  Ojos Project & Iris contributors
+// License:  GNU General Public License v3.0
 use std::fs;
-use tauri::{AppHandle, Manager};
+use tauri::AppHandle;
 use tauri_plugin_updater::UpdaterExt;
 
-/*
-#[tauri::command]
-pub fn delete_iris_data(app: AppHandle) -> Result<(), String> {
-    let os = std::env::consts::OS;
-
-    match os {
-        "windows" => {
-            let path = data_dir().unwrap().join("org.ojosproject.Iris");
-            if path.exists() {
-                std::fs::remove_dir_all(&path).map_err(|e| e.to_string())?;
-            }
-        }
-        // "macos" => {
-        //     let path = home.join("Library").join("Application Support").join("org.ojosproject.Iris");
-        //     if path.exists() {
-        //         std::fs::remove_dir_all(&path).map_err(|e| e.to_string())?;
-        //     }
-        // }
-        // "linux" => {
-        //     let path1 = home.join(".local").join("share").join("org.ojosproject.Iris");
-        //     let path2 = home.join(".config").join("org.ojosproject.Iris");
-        //     if path1.exists() {
-        //         std::fs::remove_dir_all(&path1).map_err(|e| e.to_string())?;
-        //     }
-        //     if path2.exists() {
-        //         std::fs::remove_dir_all(&path2).map_err(|e| e.to_string())?;
-        //     }
-        // }
-        _ => return Err("Unsupported OS".to_string()),
-    }
-
-    match check_update(app, true) {
-        Ok(_) => {
-            //TODO fix restart app not working in dev, works fine with msi in cargo tauri build
-            //app.restart();
-
-            //TODO placeholder, remove after fixing restart app not working in dev
-            Ok(())
-        }
-        Err(e) => Err(e.to_string()),
-    }
-}
-*/
-// we'll bring this back in a later iteration
+use crate::helpers::data_dir;
 
 #[tauri::command]
 pub fn check_update(app: AppHandle) {
@@ -66,7 +27,7 @@ pub fn check_update(app: AppHandle) {
                 )
                 .await
                 .unwrap();
-            let app_data_dir = app.path().app_data_dir().unwrap();
+            let app_data_dir = data_dir(&app);
             fs::remove_dir_all(app_data_dir).unwrap_or_else(|error| {
                 println!("Failed to remove app_data_dir: `{:?}`", error);
             });
